+++ xlwings/base_classes.py
@@ -1951,6 +1951,14 @@ class Chart:
         raise NotImplementedError()
 
     @property
+    def category_axis(self):
+        raise NotImplementedError()
+
+    @property
+    def value_axis(self):
+        raise NotImplementedError()
+
+    @property
     def plot_by(self):
         raise NotImplementedError()
 
@@ -2009,6 +2017,96 @@ class Chart:
 
     async def get_png(self):
         raise NotImplementedError("get_png() is only supported in xlwings Lite")
+
+
+class ChartAxis:
+    @property
+    def api(self):
+        raise NotImplementedError()
+
+    @property
+    def title(self):
+        raise NotImplementedError()
+
+    @title.setter
+    def title(self, value):
+        raise NotImplementedError()
+
+    @property
+    def minimum_scale(self):
+        raise NotImplementedError()
+
+    @minimum_scale.setter
+    def minimum_scale(self, value):
+        raise NotImplementedError()
+
+    @property
+    def maximum_scale(self):
+        raise NotImplementedError()
+
+    @maximum_scale.setter
+    def maximum_scale(self, value):
+        raise NotImplementedError()
+
+    @property
+    def major_unit(self):
+        raise NotImplementedError()
+
+    @major_unit.setter
+    def major_unit(self, value):
+        raise NotImplementedError()
+
+    @property
+    def number_format(self):
+        raise NotImplementedError()
+
+    @number_format.setter
+    def number_format(self, value):
+        raise NotImplementedError()
+
+    @property
+    def visible(self):
+        raise NotImplementedError()
+
+    @visible.setter
+    def visible(self, value):
+        raise NotImplementedError()
+
+    def set(
+        self,
+        *,
+        title=_UNSET,
+        minimum_scale=_UNSET,
+        maximum_scale=_UNSET,
+        major_unit=_UNSET,
+        number_format=_UNSET,
+        visible=_UNSET,
+    ):
+        raise NotImplementedError()
+
+    async def get_title(self):
+        raise NotImplementedError("get_title() is only supported in xlwings Lite")
+
+    async def get_minimum_scale(self):
+        raise NotImplementedError(
+            "get_minimum_scale() is only supported in xlwings Lite"
+        )
+
+    async def get_maximum_scale(self):
+        raise NotImplementedError(
+            "get_maximum_scale() is only supported in xlwings Lite"
+        )
+
+    async def get_major_unit(self):
+        raise NotImplementedError("get_major_unit() is only supported in xlwings Lite")
+
+    async def get_number_format(self):
+        raise NotImplementedError(
+            "get_number_format() is only supported in xlwings Lite"
+        )
+
+    async def get_visible(self):
+        raise NotImplementedError("get_visible() is only supported in xlwings Lite")
 
 
 class ChartLegend: